+++ src/test/java/org/opensearch/security/auth/http/jwt/keybyoidc/SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest.java
@@ -9,17 +9,17 @@
  * GitHub history for details.
  */
 
-package com.amazon.dlic.auth.http.jwt.keybyoidc;
+package org.opensearch.security.auth.http.jwt.keybyoidc;
 
 import java.util.HashMap;
 import java.util.List;
 
 import com.google.common.collect.ImmutableMap;
+import org.hamcrest.Matchers;
 import org.junit.Assert;
 import org.junit.Test;
 
 import org.opensearch.common.settings.Settings;
-import org.opensearch.security.auth.http.jwt.keybyoidc.HTTPJwtKeyByOpenIdConnectAuthenticator;
 import org.opensearch.security.user.AuthCredentials;
 import org.opensearch.security.util.FakeRestRequest;
 
@@ -43,7 +43,7 @@ public class SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest {
             );
 
             Assert.assertNotNull(creds);
-            assertThat(creds.getUsername(), is(TestJwts.MCCOY_SUBJECT));
+            assertThat(creds.getUsername(), Matchers.is(TestJwts.MCCOY_SUBJECT));
             assertThat(creds.getAttributes().get("attr.jwt.aud"), is(List.of(TestJwts.TEST_AUDIENCE).toString()));
             assertThat(creds.getBackendRoles().size(), is(0));
             assertThat(creds.getAttributes().size(), is(4));
@@ -93,7 +93,7 @@ public class SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest {
             );
 
             Assert.assertNotNull(creds);
-            assertThat(creds.getUsername(), is(TestJwts.MCCOY_SUBJECT));
+            assertThat(creds.getUsername(), Matchers.is(TestJwts.MCCOY_SUBJECT));
             assertThat(creds.getAttributes().get("attr.jwt.aud"), is(List.of(TestJwts.TEST_AUDIENCE).toString()));
             assertThat(creds.getBackendRoles().size(), is(0));
             assertThat(creds.getAttributes().size(), is(4));
@@ -143,7 +143,7 @@ public class SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest {
             );
 
             Assert.assertNotNull(creds);
-            assertThat(creds.getUsername(), is(TestJwts.MCCOY_SUBJECT));
+            assertThat(creds.getUsername(), Matchers.is(TestJwts.MCCOY_SUBJECT));
             assertThat(creds.getAttributes().get("attr.jwt.aud"), is(List.of(TestJwts.TEST_AUDIENCE).toString()));
             assertThat(creds.getBackendRoles().size(), is(0));
             assertThat(creds.getAttributes().size(), is(4));
@@ -171,7 +171,7 @@ public class SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest {
             );
 
             Assert.assertNotNull(creds);
-            assertThat(creds.getUsername(), is(TestJwts.MCCOY_SUBJECT));
+            assertThat(creds.getUsername(), Matchers.is(TestJwts.MCCOY_SUBJECT));
             assertThat(creds.getAttributes().get("attr.jwt.aud"), is(List.of(TestJwts.TEST_AUDIENCE).toString()));
             assertThat(creds.getBackendRoles().size(), is(0));
             assertThat(creds.getAttributes().size(), is(4));
@@ -194,7 +194,7 @@ public class SingleKeyHTTPJwtKeyByOpenIdConnectAuthenticatorTest {
             );
 
             Assert.assertNotNull(creds);
-            assertThat(creds.getUsername(), is(TestJwts.MCCOY_SUBJECT));
+            assertThat(creds.getUsername(), Matchers.is(TestJwts.MCCOY_SUBJECT));
             assertThat(creds.getAttributes().get("attr.jwt.aud"), is(List.of(TestJwts.TEST_AUDIENCE).toString()));
             assertThat(creds.getBackendRoles().size(), is(0));
             assertThat(creds.getAttributes().size(), is(4));
